chore: implement review suggestions
