+++ bciers/apps/registration/app/bceidbusiness/industry_user/contacts/page.tsx
@@ -1,6 +1,6 @@
 // 🚩 flagging that for shared routes between roles, "Page" code is a component for code maintainability
 import { ContactsSearchParams } from "@/registration/app/components/contacts/types";
-import ContactsPage from "@/registration/app/components/routes/contacts/Page";
+import ContactsPage from "@/registration/app/components/contacts/Contacts";
 
 export default async function Page({
   searchParams,
