+++ bciers/apps/registration/app/components/contacts/Contacts.tsx
@@ -1,6 +1,39 @@
 import { ContactRow, ContactsSearchParams } from "./types";
 import fetchContactsPageData from "./fetchContactsPageData";
 import ContactDataGrid from "./ContactDataGrid";
+import Note from "@bciers/components/layout/Note";
+import Link from "next/link";
+import { Button } from "@mui/material";
+
+const ExternalContactsLayout = () => {
+  return (
+    <>
+      <Note>
+        <b>Note: </b>View the contacts of your operator, i.e. people who can
+        represent the operator for GGIRCA purposes. Please keep the information
+        up to date here.
+      </Note>
+      <h2 className="text-bc-primary-blue">Contacts</h2>
+      <div className="text-right">
+        <Link href="#">
+          <Button variant="contained">Add Contact</Button>
+        </Link>
+      </div>
+    </>
+  );
+};
+
+const InternalContactsLayout = () => {
+  return (
+    <>
+      <Note>
+        <b>Note: </b>View all the contacts, which can be sorted or filtered by
+        operator here.
+      </Note>
+      <h2 className="text-bc-primary-blue">Contacts</h2>
+    </>
+  );
+};
 
 // 🧩 Main component
 export default async function Contacts({
@@ -21,8 +54,14 @@ export default async function Contacts({
 
   // Render the DataGrid component
   return (
-    <div className="mt-5">
-      <ContactDataGrid initialData={contacts} isExternalUser={isExternalUser} />
-    </div>
+    <>
+      {isExternalUser ? <ExternalContactsLayout /> : <InternalContactsLayout />}
+      <div className="mt-5">
+        <ContactDataGrid
+          initialData={contacts}
+          isExternalUser={isExternalUser}
+        />
+      </div>
+    </>
   );
 }
